requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2[binary]
python-dotenv
python-keycloak
python-jose
